pyproject.toml
[project]
name = "planertia"
version = "0.1.0"
description = ""
authors = [
    {name = "min130111",email = "min130111@naver.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]

[tool.poetry]
packages = [
    { include = "backend" },
    { include = "users" },
    { include = "tasks" },
    { include = "memos" },
    { include = "categories" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
